requirements.txt
# local package
# -e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1


# backwards compatibility
pathlib2



gensim
pandas
pyspark
kafka-python
streamlit
scikit-learn
seaborn
tensorflow
tweepy==3.9.0
pydantic
strictyaml
joblib